SCALE-SEAM (81) — the numbering ran out, and a third banner was over-claiming
Nine methods out of `client.ts` (1,436 -> 1,374) into three existing mixins:

  * *is this project ready for substantial completion, and certify it?* — turnoverReadiness,
    turnoverStatus, turnoverCertify, g704Url onto apps/web/src/api/documents.ts, beside
    closeoutSummary, which ❿ brought in under *is turnover actually closing?* ModuleRecord moved
    with them into that file's type imports and out of client.ts's.
  * *what IFC class should these generic elements be?* — ifcClassify onto apps/web/src/api/model.ts.
  * *what is the market doing to my costs?* — marketSnapshot, marketContext, marketEscalate onto
    apps/web/src/api/cost.ts. marketSnapshot is global (/market/snapshot, no project id); the other
    two apply that snapshot to one project.

THE FIRST SLICE WITH NO MARK. The ⓵–⓾ glyph range is exhausted and there is no natural successor,
so the numbering stops rather than starting a fourth visual scheme. I raised this as a decision and
it was not one: the mark is OPTIONAL in roadmapLanes.test.ts's item regex, and the roadmap item and
lane cell have carried none since ⓽. The only related guard rejects a mark that is not in MARKS —
so writing none is fine, and writing an invented glyph is what would fail.

THIRD BANNER IN THREE SLICES THAT OVER-CLAIMED. ifcClassify sat under
`// --- turnover: substantial completion (G704) + record model ---`, separated from g704Url by a
blank line, with nothing to do with substantial completion — a model question filed at whatever
banner happened to be nearest. ⓽ found three questions under "AI drafting"; ⓾ found two headers
that had declined one method for contradicting reasons; this one is the same shape again.

The two remaining banners are the largest instance yet, and are what the next slices are for:
`RACI / DACI` carries 13 methods including standardsCheck and mcpTools, and `CX-1 commissioning
loop` carries 49, among them rebarBbs and rebarCheckCage. Neither is what its banner says, and
between them they hold about a third of what is left in the file.

Two type-import moves were needed and both were caught by tsc rather than assumed: documents.ts
gained ModuleRecord, client.ts lost it (TS6196, unused). Worth noting that my own verification
command printed "TYPECHECK CLEAN" over a failing tsc on the first attempt — the `&& echo` chained
off `head`, not off `tsc`. Re-run checking the exit code directly. A false green in the check I was
using to prove there was no false green.

Verified: tsc --noEmit exit 0; full web suite 202 files / 2,056 tests; eslint clean; test_file_sizes
passes at the lowered pin; test_changelog_current, roadmapLanes and test_claude_md_gates pass.

No version bump — eighteenth follow-on under the v0.3.1143 heading.
